Revert "Revert "Merge remote-tracking branch 'origin/dev' into CardSwipeLogin""
This reverts commit f4205407ac3ce65bd0e66e924b85f6a088f3419b.

+++ src/main/java/edu/wpi/cs3733d18/teamp/ui/admin/MapBuilderEdgeFormController.java
@@ -14,10 +14,20 @@ import edu.wpi.cs3733d18.teamp.Pathfinding.Node;
 import edu.wpi.cs3733d18.teamp.ui.admin.MapBuilderController;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
 import javafx.scene.control.*;
 import javafx.scene.layout.GridPane;
+import javafx.stage.Modality;
+import javafx.stage.Stage;
+
+import java.io.IOException;
 
 public class MapBuilderEdgeFormController {
+
+    DeletePopUpController deletePopUpController;
+
     DBSystem db = DBSystem.getInstance();
 
     Boolean editFlag = false;
@@ -166,20 +176,48 @@ public class MapBuilderEdgeFormController {
      * Deletes an edge
      */
     public void deleteButtonOp(){
-        try{
+
+        // Check if deleting this edge will orphan a node
+        try {
             db.willEdgeOrphanANode(editedEdgeID);
         } catch (OrphanNodeException o) {
-            //edgeFormErrorLabel.setText("Deleting Edge will orphan Node: " + o.getNodeID());
-            //edgeFormErrorLabel.setVisible(true);
+            edgeFormErrorLabel.setText("Deleting Edge will orphan Node: " + o.getNodeID());
+            edgeFormErrorLabel.setVisible(true);
             return;
         }
+
+        // Delete Node Popup if edge can be deleted
+        Stage stage;
+        Parent root;
+        FXMLLoader loader;
+
+        stage = new Stage();
+        loader = new FXMLLoader(getClass().getResource("/FXML/general/ConfirmationPopUp.fxml"));
+
         try {
-            db.deleteEdge(editedEdgeID);
-        } catch (NodeNotFoundException | EdgeNotFoundException ne) {
-            ne.printStackTrace();
+            root = loader.load();
+        } catch (IOException ie) {
+            ie.printStackTrace();
+            return;
         }
-        mapBuilderController.updateMap();
-        mapBuilderController.addOverlay();
+
+        deletePopUpController = loader.getController();
+        deletePopUpController.StartUp(this);
+        stage.setScene(new Scene(root, 600, 150));
+        stage.initModality(Modality.APPLICATION_MODAL);
+        stage.initOwner(deleteButton.getScene().getWindow());
+        stage.showAndWait();
+
+        // If user confirms, delete edge
+        if (deletePopUpController.getChoice()) {
+            try {
+                db.deleteEdge(editedEdgeID);
+            } catch (NodeNotFoundException | EdgeNotFoundException ne) {
+                ne.printStackTrace();
+            }
+            mapBuilderController.updateMap();
+            mapBuilderController.addOverlay();
+        }
     }
 
 }
